+++ simple-marketplace/components/HeroSection.tsx
