+++ simple-marketplace/components/SignInButton.tsx
@@ -1,0 +1,9 @@
+import { MbButton } from "mintbase-ui";
+
+export const SignInButton = ({signIn}: {signIn: () => void}) => {
+
+  return (  <div className="mt-4">
+    <MbButton onClick={signIn} label="Connect NEAR Wallet" />
+  </div>
+  );
+}
