+++ simple-marketplace/config/constants.ts
@@ -1,0 +1,2 @@
+export const BINANCE_API = "https://api.binance.com/api/v3/ticker/price?symbol=NEARUSDT"
+export const MED_GAS = '300000000000000';
