+++ simple-marketplace/utils/index.ts
@@ -2,10 +2,11 @@ import { parseYactoToNear } from '../lib/numbers';
 
 export const updateThingData = ({
     data,
-
 }) => {
 
-    if (!data) {
+    const listings = data?.listings[0]
+
+    if (!listings || listings === null) {
         return {
             amountAvailable: 0,
             tokensTotal: 0,
@@ -13,19 +14,10 @@ export const updateThingData = ({
             tokenId: null,
         };
 
-    } console.log('data aa', data);
+    }
 
-    const { price, tokenId } = data.listings[0];
+    const { price, tokenId } = listings;
 
-    // const mTokensListedSaleCounter = data.simpleSaleCount.aggregate.count;
-    // const mTokensListedAuctionCounter = data.rollingAuctionCount.aggregate.count;
-    // const mTokensTotal = data.tokenCount.aggregate.count;
-    // const mPrice = ;
-
-    // const price = listThing?.price;
-    // const amountAvailable = listThing?.tokensListedSaleCounter;
-    // const tokensTotal = data.tokenCount.aggregate.count;
-    // const tokenId = listThing?.tokenId;
 
     return {
         amountAvailable: data.simpleSaleCount.aggregate.count,
